# aura-chat-service/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Aura Chat Service API
  description: |
    API REST para el servicio de chat estilo ChatGPT de Aura.
    
    ## Autenticación
    Todos los endpoints requieren autenticación JWT Bearer.
    
    ## Streaming
    Los endpoints de streaming utilizan Server-Sent Events (SSE).
    
    ## Paginación
    Los listados utilizan paginación basada en cursor.
  version: 1.0.0
  contact:
    name: Aura Team
  license:
    name: MIT
servers:
  - url: http://localhost:8002/api/v1
    description: Development server
  - url: https://api.aura.app/api/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /conversations/{conversationId}/messages:
    post:
      tags:
        - Messages
      summary: Enviar mensaje (no-streaming)
      description: |
        Envía un mensaje del usuario y obtiene la respuesta del asistente.
        La respuesta se devuelve completa una vez generada.
      operationId: sendMessage
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - name: X-Idempotency-Key
          in: header
          description: Clave de idempotencia para evitar duplicados
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
            examples:
              simple:
                summary: Mensaje simple
                value:
                  content: "¿Cómo funciona la fotosíntesis?"
              with_tools:
                summary: Con herramientas
                value:
                  content: "¿Qué clima hace en Buenos Aires?"
                  tools:
                    - type: "function"
                      function:
                        name: "get_weather"
                        description: "Obtiene el clima actual de una ciudad"
                        parameters:
                          type: "object"
                          properties:
                            location:
                              type: "string"
                              description: "Nombre de la ciudad"
                          required: ["location"]
      responses:
        '200':
          description: Mensaje enviado y respuesta generada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '502':
          $ref: '#/components/responses/LLMError'

    get:
      tags:
        - Messages
      summary: Listar mensajes
      description: Lista los mensajes de una conversación con paginación.
      operationId: listMessages
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - name: cursor
          in: query
          description: Cursor para paginación
          schema:
            type: string
        - name: limit
          in: query
          description: Cantidad de resultados por página
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: order
          in: query
          description: Orden de los mensajes
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: include_deleted
          in: query
          description: Incluir mensajes eliminados
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Lista de mensajes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /conversations/{conversationId}/messages/stream:
    post:
      tags:
        - Messages
      summary: Enviar mensaje (streaming)
      description: |
        Envía un mensaje del usuario y obtiene la respuesta del asistente en streaming.
        Utiliza Server-Sent Events (SSE) para transmitir la respuesta en tiempo real.
        
        ## Eventos SSE
        
        | Evento | Descripción |
        |--------|-------------|
        | `message_start` | Inicio del mensaje con IDs |
        | `content_delta` | Fragmento de texto |
        | `tool_call_start` | Inicio de una tool call |
        | `tool_call_delta` | Fragmento de argumentos |
        | `message_complete` | Mensaje completo con usage |
        | `error` | Error durante streaming |
        | `done` | Fin del stream |
      operationId: sendMessageStream
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - name: X-Idempotency-Key
          in: header
          description: Clave de idempotencia
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Stream de eventos SSE
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                stream:
                  summary: Ejemplo de stream
                  value: |
                    event: message_start
                    data: {"message_id": "msg-uuid", "conversation_id": "conv-uuid"}
                    
                    event: content_delta
                    data: {"delta": "La fotosíntesis es "}
                    
                    event: content_delta
                    data: {"delta": "el proceso mediante el cual..."}
                    
                    event: message_complete
                    data: {"message_id": "msg-uuid", "finish_reason": "stop", "usage": {"prompt_tokens": 50, "completion_tokens": 180, "total_tokens": 230}}
                    
                    event: done
                    data: [DONE]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
  /conversations/{conversationId}/messages/{messageId}:
    get:
      tags:
        - Messages
      summary: Obtener mensaje
      description: Obtiene los detalles de un mensaje específico.
      operationId: getMessage
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/MessageId'
      responses:
        '200':
          description: Detalle del mensaje
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Messages
      summary: Eliminar mensaje
      description: Elimina un mensaje específico (soft delete).
      operationId: deleteMessage
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/MessageId'
      responses:
        '204':
          description: Mensaje eliminado exitosamente
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /conversations/{conversationId}/messages/{messageId}/regenerate:
    post:
      tags:
        - Messages
      summary: Regenerar respuesta
      description: |
        Regenera la respuesta del asistente para un mensaje específico.
        Elimina la respuesta anterior y genera una nueva.
      operationId: regenerateMessage
      parameters:
        - $ref: '#/components/parameters/ConversationId'
        - $ref: '#/components/parameters/MessageId'
        - name: stream
          in: query
          description: Usar streaming para la respuesta
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Nueva respuesta generada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/LLMError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del servicio de autenticación
  parameters:
    ConversationId:
      name: conversationId
      in: path
      required: true
      description: ID único de la conversación (UUID)
      schema:
        type: string
        format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    MessageId:
      name: messageId
      in: path
      required: true
      description: ID único del mensaje (UUID)
      schema:
        type: string
        format: uuid
      example: "660e8400-e29b-41d4-a716-446655440001"
  schemas:
...
    SendMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 100000
          description: Contenido del mensaje
          example: "¿Cómo funciona la fotosíntesis?"
        role:
          type: string
          enum: [user, system]
          default: user
          description: Rol del mensaje
        attachments:
          type: array
          maxItems: 10
          items:
            $ref: '#/components/schemas/AttachmentRef'
          description: Referencias a adjuntos
        tools:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/ToolDefinition'
          description: Herramientas disponibles para el LLM
        metadata:
          type: object
          additionalProperties: true
          description: Metadata del mensaje
    SendMessageResponse:
      type: object
      required:
        - user_message
        - assistant_message
        - conversation
      properties:
        user_message:
          $ref: '#/components/schemas/Message'
        assistant_message:
          $ref: '#/components/schemas/Message'
        conversation:
          type: object
          properties:
            id:
              type: string
              format: uuid
            message_count:
              type: integer
            total_tokens:
              type: integer
    Message:
      type: object
      required:
        - id
        - role
        - status
        - sequence_number
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: ID único del mensaje
        role:
          type: string
          enum: [system, user, assistant, tool]
          description: Rol del mensaje
        content:
          type: string
          nullable: true
          description: Contenido del mensaje
        status:
          type: string
          enum: [draft, streaming, complete, error]
          description: Estado del mensaje
        prompt_tokens:
          type: integer
          nullable: true
          description: Tokens del prompt
        completion_tokens:
          type: integer
          nullable: true
          description: Tokens de la respuesta
        total_tokens:
          type: integer
          nullable: true
          description: Total de tokens
        model_used:
          type: string
          nullable: true
          description: Modelo que generó la respuesta
        finish_reason:
          type: string
          nullable: true
          enum: [stop, length, tool_calls, error]
          description: Razón de finalización
        tool_call_id:
          type: string
          nullable: true
          description: ID de tool call (para role=tool)
        sequence_number:
          type: integer
          description: Orden en la conversación
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
          description: Adjuntos del mensaje
        tool_calls:
          type: array
          items:
            $ref: '#/components/schemas/ToolCall'
          description: Tool calls solicitadas
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    MessageListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ==================== Attachment Schemas ====================
    AttachmentRef:
      type: object
      required:
        - file_id
        - type
      properties:
        file_id:
          type: string
          format: uuid
          description: ID del archivo subido
        type:
          type: string
          enum: [image, file, audio, video]
          description: Tipo de adjunto
    Attachment:
      type: object
      required:
        - id
        - file_name
        - file_type
        - mime_type
        - file_size
        - created_at
      properties:
        id:
          type: string
          format: uuid
        file_name:
          type: string
          description: Nombre original del archivo
        file_type:
          type: string
          enum: [image, file, audio, video]
        mime_type:
          type: string
          description: MIME type del archivo
        file_size:
          type: integer
          format: int64
          description: Tamaño en bytes
        url:
          type: string
          format: uri
          description: URL para descargar el archivo
        description:
          type: string
          nullable: true
          description: Descripción generada por AI
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time

    # ==================== Tool Call Schemas ====================
    ToolDefinition:
      type: object
      required:
        - type
        - function
      properties:
        type:
          type: string
          enum: [function]
          description: Tipo de herramienta
        function:
          $ref: '#/components/schemas/FunctionDefinition'
    FunctionDefinition:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 255
          description: Nombre de la función
        description:
          type: string
          maxLength: 1000
          description: Descripción de qué hace la función
        parameters:
          type: object
          description: JSON Schema de los parámetros
          additionalProperties: true
    ToolCall:
      type: object
      required:
        - id
        - type
        - function
      properties:
        id:
          type: string
          description: ID único de la tool call
        type:
          type: string
          enum: [function]
        function:
          type: object
          required:
            - name
            - arguments
          properties:
            name:
              type: string
              description: Nombre de la función a llamar
            arguments:
              type: string
              description: Argumentos en formato JSON string
...
    Pagination:
      type: object
      required:
        - has_more
      properties:
        next_cursor:
          type: string
          nullable: true
          description: Cursor para la siguiente página
        has_more:
          type: boolean
          description: Indica si hay más resultados
        total_count:
          type: integer
          description: Total de elementos (opcional)
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Código de error
        message:
          type: string
          description: Mensaje descriptivo del error
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorDetail'
          description: Detalles adicionales (para errores de validación)
    ValidationErrorDetail:
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
          description: Ubicación del error
        msg:
          type: string
          description: Mensaje de error
        type:
          type: string
          description: Tipo de error
  responses:
    BadRequest:
      description: Solicitud inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "BAD_REQUEST"
            message: "Invalid request body"
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "INVALID_TOKEN"
            message: "Invalid or expired token"
    Forbidden:
      description: Acceso denegado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "ACCESS_DENIED"
            message: "You don't have access to this resource"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "CONVERSATION_NOT_FOUND"
            message: "Conversation not found"
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "VALIDATION_ERROR"
            message: "Request validation failed"
            details:
              - loc: ["body", "content"]
                msg: "field required"
                type: "value_error.missing"
    RateLimitExceeded:
      description: Límite de rate excedido
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Límite de requests
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Requests restantes
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Unix timestamp cuando se resetea el límite
        Retry-After:
          schema:
            type: integer
          description: Segundos hasta poder reintentar
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "RATE_LIMIT_EXCEEDED"
            message: "Too many requests, please try again later"
    LLMError:
      description: Error del servicio LLM
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "LLM_ERROR"
            message: "Error communicating with LLM service"